Add gitignore and README files
A .gitignore file is added to ignore unnecessary files in the .idea directory. Additionally, a README.md file is also added to provide an overview of the tech challenge, the AWS resources provisioned through Terraform, and the CI/CD workflow. It also contains sections explaining Git Flow, and the deployment pipeline.